Libqaul: Fixed network topology graph closes #60
Bugfix closes #60 : Fixed network topology view: The value is a float
value. For the JSON it needs to be converted to an integer because float
values are displayed with a colon instead of a dot in some locales. This
breaks the script.
Bugfix: Network topology view: some links were not displayed.
Enhancement: If javascript is not enabled in the web browser, there is a now a warning message displayed.

+++ src/libqaul/qaullib_defines.h
@@ -192,7 +192,7 @@ struct qaul_node_msg
 {
 	union olsr_ip_addr ip;
 	union olsr_ip_addr gateway;
-	uint32_t lq;
+	float			   lq;
 };
 
 struct qaul_userhello_msg
